requirements.txt
# =============================================================================
# AlignVLA 依赖文件
# =============================================================================
# 安装命令:
#   pip install -r requirements.txt
#   pip install flash-attn --no-build-isolation  # 单独安装，需要CUDA
# =============================================================================

# -----------------------------------------------------------------------------
# 核心深度学习
# -----------------------------------------------------------------------------
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.26.0
deepspeed>=0.12.0

# -----------------------------------------------------------------------------
# LoRA和对齐训练
# -----------------------------------------------------------------------------
peft>=0.7.0
trl>=0.7.0                    # 用于DPO训练

# -----------------------------------------------------------------------------
# 数据处理
# -----------------------------------------------------------------------------
datasets>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.66.0

# -----------------------------------------------------------------------------
# 配置管理
# -----------------------------------------------------------------------------
pyyaml>=6.0
omegaconf>=2.3.0              # 支持配置文件变量引用

# -----------------------------------------------------------------------------
# nuScenes数据集
# -----------------------------------------------------------------------------
nuscenes-devkit>=1.1.10
pyquaternion>=0.9.9

# -----------------------------------------------------------------------------
# Qwen3-VL专用
# -----------------------------------------------------------------------------
qwen-vl-utils>=0.0.14

# -----------------------------------------------------------------------------
# 可视化
# -----------------------------------------------------------------------------
matplotlib>=3.7.0
opencv-python>=4.8.0
tensorboard>=2.15.0

# -----------------------------------------------------------------------------
# 其他工具
# -----------------------------------------------------------------------------
jsonlines>=4.0.0              # JSONL文件读写
scipy>=1.11.0                 # 轨迹平滑等